Refactor flask code into its own subpackage.
