+++ example/apps/simple/flask_simple_user_management.py
@@ -100,7 +100,7 @@ base_app = Flask(
     static_folder=os.path.relpath(STATIC_FOLDER, app_dir),
 )
 
-from maeser.blueprints import AppManager
+from maeser.flask_app.blueprints import AppManager
 
 # Create the AppManager class
